+++ examples/functions/sort_list_number/function.tf
@@ -1,7 +1,7 @@
 # Return the sorted list of integers:
-provider::stdlib::sort_list_string([0, 4, -10, 8])
+provider::stdlib::sort_list_number([0, 4, -10, 8])
 # result => [-10, 0, 4, 8]
 
 # Return the sorted list of floats:
-provider::stdlib::sort_list_string([9.0, 45.5, 123.4, -0.5])
+provider::stdlib::sort_list_number([9.0, 45.5, 123.4, -0.5])
 # result => [-0.5, 9.0, 45.5, 123.4]